+++ src/fileaccess.jl
@@ -92,53 +92,3 @@ function read_pdb(fname::AbstractString)
 
     return Trajectory(reshape(xyz ,3, length(atoms), length(crystal)), atoms, crystal)
 end
-
-
-
-
-function _rdf_from_file(fname::AbstractString, ur1::AbstractUnitRange,
-                     ur2::AbstractUnitRange; mindis=undef, maxdis=9.0, nbins=100)
-    try
-        t = read_trajectory(fname)
-        return compute_rdf(t, ur1, ur2, mindis=mindis, maxdis=maxdis, nbins=nbins)
-    catch
-        @warn "file $fname failed"
-    end
-        return Dict()
-end
-
-"""
-    rdf_from_files(ur1::AbstractUnitRange, ur2::AbstractUnitRange, fnames...;
-                     mindis=0.0, maxdis=9.0, nbins=100)
-"""
-function rdf_from_files(ur1::AbstractUnitRange, ur2::AbstractUnitRange,
-                        fnames...; mindis=0.0, maxdis=9.0, nbins=100)
-    dtmp = pmap( x -> _rdf_from_file(x, ur1, ur2, mindis=mindis, maxdis=maxdis, nbins=nbins) , fnames)
-
-    # Sum up results
-    data = []
-    for x in dtmp
-        if length(keys(x)) == 0
-            continue
-        end
-        push!(data,x)
-    end
-    rdf = Dict()
-    if length(data) >0
-        for (k,v) in data[1]["rdf"]
-            push!(rdf, k => deepcopy(v))
-        end
-    end
-    if length(data) > 1
-        for x in data[2:end]
-            for (k,v) in x["rdf"]
-                rdf[k] .+= v
-            end
-        end
-    end
-    for k in keys(rdf)
-        rdf[k] ./= length(data)
-    end
-    rk = collect(keys(data[1]["r"]))[1]
-    return Dict("r"=>data[1]["r"][rk] , "rdf" => rdf)
-end
